move `parenthesized` to its own module
