+++ components/salsa-macros/src/query_group.rs
@@ -1,3 +1,4 @@
+use crate::parenthesized::Parenthesized;
 use heck::CamelCase;
 use proc_macro::TokenStream;
 use proc_macro2::Span;
@@ -257,17 +258,4 @@ enum QueryStorage {
     Volatile,
     Dependencies,
     Input,
-}
-
-struct Parenthesized<T>(pub T);
-
-impl<T> syn::parse::Parse for Parenthesized<T>
-where
-    T: syn::parse::Parse,
-{
-    fn parse(input: syn::parse::ParseStream) -> syn::Result<Self> {
-        let content;
-        syn::parenthesized!(content in input);
-        content.parse::<T>().map(Parenthesized)
-    }
 }
